requirement.txt
django==1.8
markdown
django-filter
djangorestframework